@@ -3,7 +3,8 @@ Base test-case class for pwned-passwords-django.
 
 """
 
-import mock
+from unittest import mock
+
 from django.test import TestCase
 
 from pwned_passwords_django import api
@@ -14,9 +15,9 @@ class PwnedPasswordsTests(TestCase):
     Base test-case class defining some common code.
 
     """
-    sample_password = u'swordfish'
-    sample_password_prefix = u'4F571'
-    sample_password_suffix = u'81DCAADE980555F2CE6755CA425F00658BE'
+    sample_password = 'swordfish'
+    sample_password_prefix = '4F571'
+    sample_password_suffix = '81DCAADE980555F2CE6755CA425F00658BE'
     user_agent = {'User-Agent': api.USER_AGENT}
 
     def _get_mock(self, response_text=None):
