@@ -10,7 +10,6 @@ import sys
 
 import requests
 from django.conf import settings
-from django.utils.six import text_type
 
 from . import __version__
 
@@ -63,9 +62,9 @@ def pwned_password(password):
     Checks a password against the Pwned Passwords database.
 
     """
-    if not isinstance(password, text_type):
-        raise TypeError('Password values to check must be Unicode strings.')
-    password_hash = hashlib.sha1(password.encode('utf-8')).hexdigest().upper()
+    if not isinstance(password, str):
+        raise TypeError('Password values to check must be strings.')
+    password_hash = hashlib.sha1(password.encode()).hexdigest().upper()
     prefix, suffix = password_hash[:5], password_hash[5:]
     results = _get_pwned(prefix)
     if results is None:
